Add a resourceRecordUpgrader for upgrading homebrew resourses in storage
